FIR-347 ZEBI-10799 - Fix the wrong use of logical AND in place of bitwise AND operator.
Statd while adding a host to the monitor queue, calls record_addr().
record_addr() while trying to validate the IP address, intends to ensure that
the first octet is not 0. While doing so, used the logical AND instead of the
bitwise AND operator.

Review: https://jira.tegile.com:8445/cru/ZB35X-299 (Sai, Santosh)

Testing: Acquired a lock on a file and ensured that statd records the client
addr and name in /var/statmon/sm.