BSR-5533 smartpqi DDI entry point cleanups
While here we've converted a few of the structure definitions to
C99 style which is more resistent to coding errors than the legacy
C89 position based style.  Note that getinfo() will be populated
by the SCSA framework automatically with a suitable entry point.

The use of scsi_hba_ioctl also permits a SCSI based attachment
point to be used with cfgadm.