Let SMB server use smbcrypt kmod/lib (modifications)
The new kmod (smbcrypt) and new lib (libfksmbcrypt) are part of
the client because the server packages depend on the client pkgs.
The server now uses that kmod and library.

The smb31_preauth.c and fksmb_preauth_pkcs.c could move too,
depending on whether that's needed in the client.