BSR-10840 WPTS SMB 3.1.1 tests fail when encryption enabled
We should not send a SMB2_ENCRYPTION_CAPS unless the request
contained one.

Don't gratuitously encrypt responses.  Only encrypt when the
request was encrypted, and demand an encrypted request when
the rules say it should be encrypted.

Don't set tree connect response "encrypted" flag unless the
tree is set to require encryption.