14306 Can't get SID for ID=0 type=0, status=-9977 (partial)
Note, this is by NO means a complete fix, and might not fix this at all.
Rather, the point of this is to (a) fix the console noise, (b) avoid noise
when we know the caller doesn't care, and and (c) get better information
about what the mapping request was that failed.

If people could try this change, that might help get us to a "real" fix.

Change-Id: I4f5e0ff1d10447a1287098d8107ca1bb5f535d06