BSR-14422 SMB2 async. responses could be more like Windows
We observe a client (Linux 6.1 cifs mount) that has problems with SMB2 interim
responses.  This client works OK with windows, but not our SMB server.
Our SMB2 server sends an interim response for the command that needs to block,
and also sends interim responses for any following commands in that compound
(as the SMB2 spec. seems to suggest).
In our observations of Windows server behavior, it appears that the command
that needs to block gets an interim response, but that interim response is
sent as a "singleton", not as a compound like the request.
Do like Windows does to keep such clients happy.